Move ConnectionsStore to stores dir
